🌳 ext.lib.peerjs (webrtc) (#167)
* update callsite

* PeerModel events

* refactor

* Peer data sending

* refactor (rename file)

* UntilObservable taking deep input array

* test: ensure deep

* rename extension

* refactor

* delete files

* refactor

* refactor

* fix call sites

* rename type

* type renames

* fix call sites

* rename Peer → PeerJS

* move module to own folder

* refactor (module out)

* refactor

* fix

* connection ID returned

* root init method

* comments

* peer id enhancements

* Update PeerJs.ts

* format params

* refactor to use model

* move types

* expose { toObject } helpers

* clean

* rename: Id.is → Is.Is

* custon ID length

* ID of specified length (custom)

* Model.action helper

* refactor: peer-id

* Update ui.PeerLabel.tsx

* sample tweaks

* folder rename (samples)

* return type

* Update -TEST.integration.ts

* Update Model.-TEST.v.ts

* list redraw (controller)

* remove total from dev-spec (so that redraw in list controller tests well)

* connecting adding new list item

* refactor sample

* refactor

* refactor

* Update -TEST.Uri.v.ts

* refactor: rename "local" → "self"

* sample

* rename: rx.distinctUntilChanged → distinctWhile (shorter)

* Connect to network (interaction sequence)

* tweaks

* remote connected icon

* remote - copy to clipboard

* Close Connection (UX interaction sequence)

* List event: → selected$

* Update -SPEC.tsx

* <Info> → Peer row

* template tweak

* types

* tweaks

* Update ui.PeerCard.tsx

* rename types

* util → rxjs → throttleAnimationFrame

* ListArray helper - typed →  fill, map, remove methods

* remove command

* updates for "remove from list"

* remove - only index in message param

* command return self (chaining)

* Update use.ListController.tsx

* Update -SPEC.tsx

* simplify

* prefix → "self"

* close icon

* store connection id

* Update ui.PeerCard.tsx

* Update -SPEC.tsx

* self icon (animated on open)

* delete/close connection

* remove item - tweaks

* prevent duplicate peer connections

* close

* render connected component (right)

* build

* french translation sample (video → ML)

* clean / docs

* Media toolbar (WIP)

* Delete ui.Connected.tsx

* type rename

* Update -SPEC.tsx

* CMD + Escape for "up to root" command (double)

* clear pasted on escape

* clean

* Update Root.tsx

* refactor

* refactor

* refactor

* List → item specific event hookups

* Purge - response data

* purge (self → Delete)

* Remote - <ActionComplete> object

* rename folder

* spec

* Media tests

* connect to media (video)

* close media connection

* close child media connections

* return { conn } on data

* rename API

* getUserScreen

* screen sharing (setup within model and rough spec/sample UI)

* rename file

* release media streams on connection close

* Connect (via MediaToolbar component)

* rename type

* connector buttons

* update refs

* update refs

* Update package.json

* type files: .mts → .ts

* build

* stub in <Bitext> component

* rename folder

* refactor

* move types

* refactor

* sample editors

* build

* clean

* update ESM template: "sample" → "sample.01"

* more elegant

* PropList → take falsey array

* Update -Sample.Builder.tsx

* refactor: PeerModel.get (function exposes properties)

* Update ui.PeerCard.tsx

* refactor: stream memory manage → utility module

* Update PeerModel.ts

* refactor

* get connections by peer

* more redraws

* show remotes

* extend info field type

* Is.kind<conn>

* call sites: using Is.media<kind>

* <Info> panel showing remote peers and conns

* entry param

* <Connector> - component entry

* render peer <Connector> within <Info> spec

* <Antenna> icon on connected peer

* keyboard behavior: grab focus

* behavior: focusOnLoad

* Update -SPEC.tsx

* build

* peerjs → <Connector> placed within <App> footer

* Update -SPEC.tsx

* Refactor

* userAgent

* refactor: stream retrieval returning stream AND/OR error

* refactor

* fix: multiple screenshare attempts

* disconnect media (on toggle button)

* stub in <AvatarTray>

* fire onReady

* <AvatarTray>

* fix which stream shows (supress flash between Self → Remote)

* no render while not ready

* export <AvatarTray>

* render Avatar tray

* build

* item render fixes: null | undefined values

* null render tests

* default size

* fix call-site

* PatchState: Is.type (typename field)

* move "type" field up to sys.types (ImmutableRef)

* Update -SPEC.tsx

* move to own file

* Model.Item, Model.List - type param for state model init

* Model.Is.type ← tester

* Model type names (list, remote, self) and Model.Is flag checker

* Model: item selection$ observable (with tests)

* Edit commands

* control edit via message commands

* cmd: → "edited"

* fire connection kind through bus

* create list state from Peer (within hook)

* Update use.ListState.ts

* LabelItem editor: → Cancel [↑↓] arrow keys

* Update types.ts

* Update ui.Label.tsx

* Update Renderer.Self.tsx

* paste → connect → disconnect

* fix close behavior (and trigger from button click)

* prefix color ("self")

* LabelItem → ensure focus on mouse-down

* export <Video>

* Sample: Video (full screen)

* Clean up <Video>

* stub in App.00

* stubbed in peer → video display

* rough video display (peers)

* deploy with new <Video> peer UI

* move to top

* LabelItem renderer return <null, undefined> behavior mods

* update call sites for renderer

* Update -TEST.v.ts

* build

* hash sample

* UserAgent.current → lazily evaluated, cached response

* OS aware clipboard flags (is.copy, is.cut, is.paste)

* fire clipboard events from keyboard event flags

* build

* refactor: <dev.PeerCard>

* tmp rename

* rename: → .dev →  Dev.PeerCard

* refactor

* adjust call-site (WIP)

* update refs

* size slider

* move to archive

"@web3-storage/w3up-client" breaking build.
Dubious state of stability.

* remove refs

* build

* tweak

* FIX: added to MD → HTML sanitizer schema: "target" and "rel" attributes

* refactor dev card

* tweak

* event api: rename "action.on" →  "action.kind"

* List Item: "key.on" handler API

* fix call-sites

* Peer - connection retrieval API

* debug printing to console (CMD + P)

* on fire "onReady" once

* <Input> refactorings around useImperativeHandle

* visible in tree (debug)

* remove return value (non-chainable)

Makes interface more re-usable / composable

* <VirtualList> using imperative handle forwarding

* empty message

* LabelList - Model → get taking "First" and "Last" and Index/Id param input

* refactor

* <Connector> ref API

* export <UI.Info>

* refactor selection streams

* refactor

* refactor

* fix build

* update refs

* mute self

* version

+++ code/compiler/builder/Builder/Builder.build.mts
@@ -21,27 +21,29 @@ export async function build(
   dir = fs.resolve(dir);
   const relativeDir = dir.substring(Paths.rootDir.length);
 
-  // Pre-build.
-  await Template.ensureBaseline(dir);
-  if (syncDeps) {
-    await Dependencies.syncVersions({ filter: (dir) => dir === relativeDir });
+  if (true) {
+    // Pre-build.
+    await Template.ensureBaseline(dir);
+    if (syncDeps) {
+      await Dependencies.syncVersions({ filter: (dir) => dir === relativeDir });
+    }
+
+    // - Typescript (lint).
+    const tsBuildOutput = await Typescript.build(dir, { exitOnError, silent });
+    if (!tsBuildOutput.ok) return tsBuildOutput;
+
+    // - ESM bundling.
+    const viteBuildOutput = await Vite.build(dir, { silent });
+    if (!viteBuildOutput.ok) return viteBuildOutput;
+
+    // Post build.
+    await fs.remove(fs.join(dir, Paths.types.dist));
+    await fs.move(fs.join(dir, Paths.types.dirname), fs.join(dir, Paths.types.dist));
+    await Package.generate(dir);
+    await Vite.deleteBuildManifests(dir);
   }
 
-  // - Typescript (lint).
-  const tsBuildOutput = await Typescript.build(dir, { exitOnError, silent });
-  if (!tsBuildOutput.ok) return tsBuildOutput;
-
-  // - ESM bundling.
-  const viteBuildOutput = await Vite.build(dir, { silent });
-  if (!viteBuildOutput.ok) return viteBuildOutput;
-
-  // Post build.
-  await fs.remove(fs.join(dir, Paths.types.dist));
-  await fs.move(fs.join(dir, Paths.types.dirname), fs.join(dir, Paths.types.dist));
-  await Package.generate(dir);
-  await Vite.deleteBuildManifests(dir);
-
-  // Finish up.
+  // Write output.
   if (!silent) {
     const typedir = `/${Paths.types.dirname}/`;
     const size = await Util.folderSize(fs.join(dir, 'dist'));
@@ -56,8 +58,16 @@ export async function build(
     const pkg = await Util.importPkgMeta(dir);
     const moduleName = pkg?.name || 'unknown';
 
+    const version = pkg?.version ?? '';
+    const major = version?.split('.')[0];
+    const minor = version?.split('.')[1];
+    const patch = version?.split('.')[2];
+    const versionDisplay =
+      patch && version !== '0.0.0' ? pc.green(`${major}.${minor}.${patch}`) : version;
+
     const table = LogTable();
     const indent = ' '.repeat(6);
+    table.push([indent, 'version', `  ${versionDisplay}`, '']);
     table.push([indent, 'bundle', `  ${bundle.length} files  `, bundleSize]);
     table.push([indent, 'types.d', `  ${types.length} files  `, typesSize]);
     table.push([indent, '', '', '-'.repeat(10)]);
@@ -69,5 +79,6 @@ export async function build(
     console.info('');
   }
 
-  return { ok: true, errorCode: 0 };
+  // Finish up.
+  return { ok: true, errorCode: 0 } as const;
 }
